+++ assets/MagicPage-nm7AZFTf.js
@@ -1,1 +1,0 @@
-import{l as e}from"./index-BFDj18tu.js";import{t}from"./ProductPage-6La0gYR8.js";import{t as n}from"./products-B8C8eSos.js";var r=e();function i(){return(0,r.jsx)(t,{product:n(`magic`),backTo:`/vedasleep`,transitionId:`product-vedasleep-magic`})}export{i as default};